requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
sqlmodel==0.0.24
asyncpg==0.30.0
azure-core==1.34.0
azure-identity==1.22.0
azure-keyvault-secrets==4.9.0
botbuilder-azure==4.16.2
azure-storage-blob==12.7.0
botbuilder-core==4.16.2
botbuilder-dialogs==4.16.2
gunicorn==23.0.0
pydantic==2.11.4
sqlalchemy==2.0.40
aiosqlite==0.21.0
timescaledb==0.0.4
httpx==0.28.1
aiohttp==3.9.5
msrest==0.7.1